git->convert, GitCommand->PatchToBranch
