--- src/Events/GitCommand/DogitEvent.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace dogit\Events\GitCommand;
-
-abstract class DogitEvent
-{
-}
