--- src/Events/GitCommand/FilterByResponseEvent.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace dogit\Events\GitCommand;
-
-final class FilterByResponseEvent extends AbstractFilterEvent
-{
-}
