+++ src/Events/PatchToBranch/DogitEvent.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace dogit\Events\PatchToBranch;
+
+abstract class DogitEvent
+{
+}
