+++ src/Events/PatchToBranch/FilterByResponseEvent.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace dogit\Events\PatchToBranch;
+
+final class FilterByResponseEvent extends AbstractFilterEvent
+{
+}
